+++ plugins/source/azure/resources/services/cosmos/database_accounts.go
@@ -14,7 +14,7 @@ func DatabaseAccounts() *schema.Table {
 		Name:                 "azure_cosmos_database_accounts",
 		Resolver:             fetchDatabaseAccounts,
 		PostResourceResolver: client.LowercaseIDResolver,
-		Description:          "https://learn.microsoft.com/en-us/rest/api/cosmos-db-resource-provider/2022-05-15/database-accounts/list?tabs=HTTP#databaseaccountgetresults",
+		Description:          "https://learn.microsoft.com/en-us/rest/api/cosmos-db-resource-provider/database-accounts/get?view=rest-cosmos-db-resource-provider-2023-11-15&tabs=HTTP#databaseaccountgetresults",
 		Multiplex:            client.SubscriptionMultiplexRegisteredNamespace("azure_cosmos_database_accounts", client.Namespacemicrosoft_documentdb),
 		Transform:            transformers.TransformWithStruct(&armcosmos.DatabaseAccountGetResults{}, transformers.WithPrimaryKeys("ID")),
 		Columns:              schema.ColumnList{client.SubscriptionID},